Split registration form into component
