+++ apps/ccf-rui/src/app/app.component.ts
@@ -1,4 +1,5 @@
 import {
+  booleanAttribute,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
@@ -9,18 +10,37 @@ import {
   OnDestroy,
   OnInit,
 } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { GlobalConfigState, TrackingPopupComponent } from 'ccf-shared';
 import { ConsentService } from 'ccf-shared/analytics';
-import { combineLatest, ReplaySubject, Subscription } from 'rxjs';
-
 import { GoogleAnalyticsService } from 'ngx-google-analytics';
+import {
+  combineLatest,
+  filter,
+  fromEvent,
+  map,
+  ReplaySubject,
+  startWith,
+  Subscription,
+  switchAll,
+  take,
+  throttleTime,
+} from 'rxjs';
+import { EMPTY_SUBSCRIPTION } from 'rxjs/internal/Subscription';
+
 import { GlobalConfig } from './core/services/config/config';
 import { ThemingService } from './core/services/theming/theming.service';
 import { ModelState, ViewSide, ViewType } from './core/store/model/model.state';
 import { PageState } from './core/store/page/page.state';
-import { Side } from './modules/content/stage-nav/stage-nav.component';
 import { RegistrationState } from './core/store/registration/registration.state';
+import { Side } from './modules/content/stage-nav/stage-nav.component';
+import {
+  DEFAULT_SCREEN_SIZE_NOTICE_STORAGE_KEY,
+  SCREEN_SIZE_NOTICE_MAX_HEIGHT,
+  SCREEN_SIZE_NOTICE_MAX_WIDTH,
+  ScreenSizeNoticeComponent,
+} from './modules/screen-size-notice/screen-size-notice.component';
 
 export interface User {
   firstName: string;
@@ -34,6 +54,35 @@ interface AppOptions extends GlobalConfig {
   logoTooltip?: string;
   view?: ViewType;
   viewSide?: ViewSide;
+}
+
+export function openScreenSizeNotice(dialog: MatDialog): Subscription {
+  const initialStorageValue = booleanAttribute(localStorage.getItem(DEFAULT_SCREEN_SIZE_NOTICE_STORAGE_KEY));
+  if (initialStorageValue) {
+    return EMPTY_SUBSCRIPTION;
+  }
+  const testScreenSize = () =>
+    window.innerWidth < SCREEN_SIZE_NOTICE_MAX_WIDTH || window.innerHeight < SCREEN_SIZE_NOTICE_MAX_HEIGHT;
+  const afterClosed$ = fromEvent(window, 'resize').pipe(
+    throttleTime(50),
+    startWith({}),
+    filter(testScreenSize),
+    take(1),
+    map(() => {
+      return dialog.open(ScreenSizeNoticeComponent, {
+        panelClass: 'screen-size-notice-panel',
+        width: '456px',
+        disableClose: true,
+        closeOnNavigation: false,
+      });
+    }),
+    map((ref) => ref.afterClosed()),
+    switchAll(),
+  );
+
+  return afterClosed$.subscribe(() => {
+    localStorage.setItem(DEFAULT_SCREEN_SIZE_NOTICE_STORAGE_KEY, 'true');
+  });
 }
 
 /**
@@ -89,6 +138,7 @@ export class AppComponent implements OnDestroy, OnInit {
     readonly consentService: ConsentService,
     readonly snackbar: MatSnackBar,
     readonly theming: ThemingService,
+    public dialog: MatDialog,
     el: ElementRef<unknown>,
     injector: Injector,
     private readonly globalConfig: GlobalConfigState<AppOptions>,
