+++ apps/ccf-rui/src/app/modules/registration-modal/registration-form/registration-form.component.scss
@@ -1,0 +1,91 @@
+form {
+  display: flex;
+  flex-direction: column;
+
+  .subheader {
+    font: var(--sys-title-small);
+    color: var(--sys-primary);
+    margin-bottom: 1rem;
+  }
+
+  .subheader.donor {
+    grid-area: subheader-donor;
+  }
+
+  .field.firstname {
+    grid-area: field-firstname;
+  }
+
+  .field.middlename {
+    grid-area: field-middlename;
+  }
+
+  .field.lastname {
+    grid-area: field-lastname;
+  }
+
+  .field.orcid {
+    grid-area: field-orcid;
+  }
+
+  .field.email {
+    grid-area: field-email;
+  }
+
+  .field.sex {
+    grid-area: field-sex;
+  }
+
+  .field.organ {
+    grid-area: field-organ;
+  }
+
+  .field.consortium {
+    grid-area: field-consortium;
+  }
+
+  .field.doi {
+    grid-area: field-doi;
+  }
+
+  .author-fields {
+    display: grid;
+    column-gap: 1.5rem;
+    row-gap: 1.5rem;
+    grid-template-areas:
+      'field-firstname field-middlename'
+      'field-lastname field-orcid'
+      'field-email field-email';
+    margin-bottom: 2rem;
+
+    @media (max-width: 759px) {
+      grid-template-areas:
+        'field-firstname'
+        'field-middlename'
+        'field-lastname'
+        'field-orcid'
+        'field-email';
+    }
+  }
+
+  .donor-fields {
+    display: grid;
+    column-gap: 1.5rem;
+    row-gap: 1.5rem;
+    grid-template-columns: 8rem 12.5rem auto;
+    grid-template-areas:
+      'field-sex field-organ field-consortium'
+      'field-doi field-doi field-doi';
+
+    @media (max-width: 759px) {
+      grid-template-areas:
+        'field-sex field-organ field-organ'
+        'field-consortium field-consortium field-consortium'
+        'field-doi field-doi field-doi';
+    }
+  }
+
+  mat-error {
+    height: 0rem;
+  }
+}
