+++ apps/ccf-rui/src/app/modules/registration-modal/registration-modal/registration-modal.component.spec.ts
@@ -97,12 +97,6 @@ describe('RegistrationModalComponent', () => {
       });
   });
 
-  // it('should open the dialog when the openDialog() method is called', async () => {
-  //   const { instance, inject } = await shallow.render();
-  //   instance.openDialog();
-  //   expect(inject(MatDialog).open).toHaveBeenCalled();
-  // });
-
   it('should open the dialog on init if there is no user or organ', async () => {
     const { instance } = await shallow.render();
     const spy = spyOn(instance, 'openDialog');
