+++ apps/ccf-rui/src/themes/common/_theming.scss
@@ -33,14 +33,10 @@
 @mixin ccf-theme($theme) {
   @include mat.core-theme($theme);
 
-  // @include mat.button-theme($theme);
   @include mat.checkbox-theme($theme);
   @include mat.expansion-theme($theme);
-  // @include mat.form-field-theme($theme);
   @include mat.icon-theme($theme);
-  // @include mat.input-theme($theme);
   @include mat.radio-theme($theme);
-  // @include mat.select-theme($theme);
   @include mat.slide-toggle-theme($theme);
   @include mat.toolbar-theme($theme);
   @include mat.tooltip-theme($theme);
